+++ example/lib/global_error_data.dart
@@ -6,6 +6,6 @@ class GlobalErrorData extends ErrorData {
     required super.title,
     required super.exception,
     super.controllerSource,
-    super.onError,
+    super.actions,
   });
 }