+++ lib/src/exceptions/app_exception.dart
@@ -1,5 +1,7 @@
+import 'package:coder_matthews_extensions/coder_matthews_extensions.dart';
+
 class AppException implements Exception {
   final String title, message;
-  final void Function()? callback;
-  AppException(this.title, this.message, [this.callback]);
+  final List<ErrorDataAction> actions;
+  AppException(this.title, this.message, {this.actions = const []});
 }